+++ app/_layout.tsx
@@ -7,8 +7,7 @@ import { GluestackUIProvider } from "@/components/ui/gluestack-ui-provider";
 import "@/global.css";
 import { DUMMY_PET_LIST } from "@/constants/dummyPetList";
 import { usePetStore } from "@/store/usePetStore";
-import { getFirebaseMessaging } from "@/util/firebaseMessage";
-import { initFCM } from "@/util/initFCM";
+import { getFirebaseMessaging, initFCM } from "@/util/notification";
 import notifee from "@notifee/react-native";
 import { onMessage } from "@react-native-firebase/messaging";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
