+++ components/board/HomeDashBoard/WalkScoreBoard.tsx
@@ -1,7 +1,7 @@
 import WalkScoreSkeleton from "@/components/skeleton/WalkScoreSkeleton";
 import { useLocationPermission } from "@/hooks/use-location-permission";
 import { useWeatherStore } from "@/store/useWeatherStore";
-import { getDustLevel } from "@/util/getDustLevel";
+import { getDustLevel } from "@/util/weather";
 import { Ionicons } from "@expo/vector-icons";
 import * as Location from "expo-location";
 import React, { useEffect, useRef, useState } from "react";
