+++ util/mbti/getResultMbti.ts
