+++ util/run/index.ts
@@ -1,0 +1,2 @@
+export * from "./calcPace";
+export * from "./formatTime";
